+++ src/main/java/com/billhu/ecommercesideproject/dao/mapper/UserMapper.java
@@ -16,7 +16,7 @@ public interface UserMapper {
      * @SelectKey insert date 查詢db自動生成的id  ， keyColumn =>對應db 欄位  keyProperty => 對應 entity 參數  resultType=> 回傳type before=>執行insert 前或後執行 statement=>要執行得sql語法
      *參考 https://blog.csdn.net/Wangdiankun/article/details/106461414
      */
-    @SelectKey( keyColumn = "user_id",keyProperty = "userId",resultType = long.class,before = false ,statement = "SELECT LAST_INSERT_ID()")
+    @SelectKey( keyColumn = "user_id",keyProperty = "userId",resultType = Integer.class,before = false ,statement = "SELECT LAST_INSERT_ID()")
     @Insert("INSERT INTO users(user_mail,user_password,user_type,create_time,modify_time) " +
             " VALUES (#{userMail},#{userPassword},#{userType},#{createTime},#{modifyTime} ) ")
     public void create (UserEntity user);
